feat: Refactor countdown display and add new font
Refactors the `CountdownDisplay` component to extract time unit rendering into a new `TimeUnitGroup` component. This improves code organization and reusability.

Also adds the 'Intel One Mono' font to the project, enhancing the cyberpunk aesthetic. The `RollingDigit` component is updated to no longer accept a `label` prop, as this is now handled by `TimeUnitGroup`. The `tick` state in `CountdownDisplay` was removed as it was unused.